Add ability to change roles from people list

+++ app/components/user-list-item.js
@@ -2,40 +2,55 @@ import Component from '@ember/component';
 import { inject as service } from '@ember/service';
 import { get, set } from '@ember/object';
 import { alias } from '@ember/object/computed';
+import RSVP from 'rsvp';
 
 export default Component.extend({
   classNames: ['user-list-item'],
   tagName: 'li',
-  showApprove: false,
-  showDeny: false,
 
   flashMessages: service(),
 
   projectUser: null,
+  showRoleModal: false,
   user: null,
 
   project: alias('projectUser.project'),
 
   actions: {
     approve(projectUser) {
-      set(projectUser, 'role', 'contributor');
+      let confirmed = window.confirm('Are you sure you want to approve their membership?');
+      if (confirmed) {
+        set(projectUser, 'role', 'contributor');
+        return projectUser.save().then(() => {
+          this._flashSuccess('Membership approved');
+        });
+      } else {
+        return RSVP.reject();
+      }
+    },
+
+    changeRole(projectUser, role) {
+      let currentRole = get(projectUser, 'role');
+      if (role === currentRole) {
+        set(this, 'showRoleModal', false);
+        return RSVP.resolve();
+      }
+
+      set(projectUser, 'role', role);
       return projectUser.save().then(() => {
-        this._flashSuccess('Membership approved');
+        this._flashSuccess(`Role changed to ${role}`);
       });
     },
 
     deny(projectUser) {
-      return projectUser.destroyRecord().then(() => {
-        this._flashSuccess('Membership denied');
-      });
-    },
-
-    showApprove() {
-      set(this, 'showApprove', true);
-    },
-
-    showDeny() {
-      set(this, 'showDeny', true);
+      let confirmed = window.confirm('Are you sure you want to deny their membership?');
+      if (confirmed) {
+        return projectUser.destroyRecord().then(() => {
+          this._flashSuccess('Membership denied');
+        });
+      } else {
+        return RSVP.reject();
+      }
     }
   },
